rag-debugger/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
faiss-cpu==1.7.4
GitPython==3.1.40
pytest==7.4.3
openai==1.6.0
httpx<0.28.0
python-dotenv==1.0.0
tiktoken==0.5.2
numpy<2
